# Sparse: faster CSR/CSC setitem with change in sparsity