@pv 2015-10-10
Closing (stalled & probably a different approach is better).